Adding check for deleting only work objects
git-svn-id: svn://oss.beweb.dbc.dk/sync_search/opensearch/trunk@7797 51cac136-f83d-432d-bb6f-125e83a60a2d

+++ src/dk/dbc/opensearch/common/fedora/FedoraMain.java
@@ -282,10 +282,13 @@ public class FedoraMain
                 while ( ( line = input.readLine()) != null )
                 {
                     String[] work_mani = line.split( " " );
-                    String pid = work_mani[ 0 ];
+                    String pid = work_mani[ 1 ];
                     for ( int i = 1; i < work_mani.length; i++ )
                     {
-                        objectRepository.deleteObject( pid, "Deleting from FedoraAuxiliaryMain" );
+                        if ( pid.startsWith( "work" ) )
+                        {
+                            objectRepository.deleteObject( pid, "Deleting from FedoraAuxiliaryMain" );
+                        }
                     }
                 }
             }
